Move the TimesFM3 PyTorch backend into a torch/ subpackage
Relocates the existing src/timesfm3 modules into src/timesfm3/torch/ to make room for
additional backends. Relative imports within the package are unchanged. The top-level
timesfm3 package lazily re-exports the PyTorch API (`from timesfm3 import TimesFM3Forecaster`,
etc.), so existing imports keep working and `import timesfm3` no longer requires torch.
