+++ src/timesfm3/torch/configs.py
