+++ src/timesfm3/torch/dense.py
