+++ src/timesfm3/torch/model.py
