+++ src/timesfm3/torch/util.py
